pyproject.toml
[project]
name = "goatvaultserver"
version = "0.1.0"
description = "FastAPI with MongoDB"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.0",
    "taskipy>=1.14.1",
    "ty>=0.0.1a22",
]

[tool.taskipy.tasks]
dev = "uvicorn app.main:app --reload"
server = "uvicorn app.main:app --host 0.0.0.0 --port 8000"
test = "pytest --cov=app --cov-report=term --cov-report=html --cov-report=xml"
lint = "ruff check"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-v"
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
branch = true
relative_files = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E","W","F","I","B","C4","UP","ARG","SIM","FAST001","FAST002","FAST003"]
ignore = ["E501"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.pylint]
max-args = 15
max-branches = 20
max-returns = 10
max-statements = 80

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"
